ZXP-116 Rename UserCompany.ts and UserStudent.ts and User.ts
