+++ front-angular/src/models/UserCompany.ts
