+++ front-angular/src/models/index.ts
@@ -1,7 +1,7 @@
 
-export * from './user';
-export * from './userCompany';
-export * from './userStudent';
+export * from './User';
+export * from './UserCompany';
+export * from './UserStudent';
 export * from './Company';
 export * from './Avis';
 export * from './Offer';
